+++ PathologyLabs.Domain/Core/Report.cs
@@ -5,7 +5,7 @@
         public PathologyTest Test { get; set; }
 
         public Patient Patient { get; set; }
-        public int PatientId { get; set; }
+        public long PatientId { get; set; }
 
         public string Comment { get; set; }
     }
